task 3 from project 0x10 is done, int to char casted, double to float casted

+++ 0x10-variadic_functions/3-print_all.c
@@ -26,7 +26,7 @@ void print_all(const char * const format, ...)
 				printf("%c", (char)va_arg(ags, int));
 				break;
 			case 'f':
-				printf("%f", va_arg(ags, double));
+				printf("%f", (float)va_arg(ags, double));
 				break;
 			case 's':
 				str = va_arg(ags, char *);
